requirements.txt
aerich==0.6.2
aiomysql==0.0.22
aiosqlite==0.17.0
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.0
click==8.0.4
ddlparse==1.10.0
dictdiffer==0.9.0
fastapi==0.75.0
h11==0.13.0
httptools==0.4.0
idna==3.3
iso8601==0.1.16
jedi==0.18.1
parso==0.8.3
prompt-toolkit==3.0.28
ptpython==3.0.20
pydantic==1.9.0
Pygments==2.11.2
PyMySQL==0.9.3
pyparsing==3.0.7
pypika-tortoise==0.1.3
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
tomlkit==0.10.0
tortoise-cli==0.1.1
tortoise-orm==0.18.1
typing_extensions==4.1.1
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.1
wcwidth==0.2.5
websockets==10.2
